+++ contracts/reflect/src/msg.rs
@@ -1,7 +1,7 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-use cosmwasm_std::{Binary, CosmosMsg, HumanAddr, QueryRequest};
+use cosmwasm_std::{Binary, CosmosMsg, CustomQuery, HumanAddr};
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct InitMsg {}
@@ -17,7 +17,7 @@ pub enum HandleMsg {
 #[serde(rename_all = "snake_case")]
 pub enum QueryMsg {
     Owner {},
-    // this will call out to CustomQuery::Capitalize
+    // this will call out to SpecialQuery::Capitalize
     ReflectCustom { text: String },
 }
 
@@ -43,22 +43,17 @@ impl Into<CosmosMsg<CustomMsg>> for CustomMsg {
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-/// CustomQuery is an override of QueryRequest::Custom to show this works and can be extended in the contract
-pub enum CustomQuery {
+/// An implementation of QueryRequest::Custom to show this works and can be extended in the contract
+pub enum SpecialQuery {
     Ping {},
     Capital { text: String },
 }
 
-// TODO: do we want to standardize this somehow for all?
-impl Into<QueryRequest<CustomQuery>> for CustomQuery {
-    fn into(self) -> QueryRequest<CustomQuery> {
-        QueryRequest::Custom(self)
-    }
-}
+impl CustomQuery for SpecialQuery {}
 
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 #[serde(rename_all = "snake_case")]
-// All return values of CustomQuery are CustomResponse
-pub struct CustomResponse {
+// All return values of SpecialQuery are SpecialResponse
+pub struct SpecialResponse {
     pub msg: String,
 }