+++ contracts/reflect/src/testing.rs
@@ -1,4 +1,4 @@
-use crate::msg::{CustomQuery, CustomResponse};
+use crate::msg::{SpecialQuery, SpecialResponse};
 
 use cosmwasm_std::testing::{MockApi, MockQuerier, MockStorage, MOCK_CONTRACT_ADDR};
 use cosmwasm_std::{to_binary, Binary, Coin, ContractResult, Extern, HumanAddr, SystemResult};
@@ -8,9 +8,9 @@ use cosmwasm_std::{to_binary, Binary, Coin, ContractResult, Extern, HumanAddr, S
 pub fn mock_dependencies_with_custom_querier(
     canonical_length: usize,
     contract_balance: &[Coin],
-) -> Extern<MockStorage, MockApi, MockQuerier<CustomQuery>> {
+) -> Extern<MockStorage, MockApi, MockQuerier<SpecialQuery>> {
     let contract_addr = HumanAddr::from(MOCK_CONTRACT_ADDR);
-    let custom_querier: MockQuerier<CustomQuery> =
+    let custom_querier: MockQuerier<SpecialQuery> =
         MockQuerier::new(&[(&contract_addr, contract_balance)])
             .with_custom_handler(|query| SystemResult::Ok(custom_query_execute(&query)));
     Extern {
@@ -20,12 +20,12 @@ pub fn mock_dependencies_with_custom_querier(
     }
 }
 
-pub fn custom_query_execute(query: &CustomQuery) -> ContractResult<Binary> {
+pub fn custom_query_execute(query: &SpecialQuery) -> ContractResult<Binary> {
     let msg = match query {
-        CustomQuery::Ping {} => "pong".to_string(),
-        CustomQuery::Capital { text } => text.to_uppercase(),
+        SpecialQuery::Ping {} => "pong".to_string(),
+        SpecialQuery::Capital { text } => text.to_uppercase(),
     };
-    to_binary(&CustomResponse { msg }).into()
+    to_binary(&SpecialResponse { msg }).into()
 }
 
 #[cfg(test)]
@@ -35,29 +35,29 @@ mod test {
 
     #[test]
     fn custom_query_execute_ping() {
-        let res = custom_query_execute(&CustomQuery::Ping {}).unwrap();
-        let response: CustomResponse = from_binary(&res).unwrap();
+        let res = custom_query_execute(&SpecialQuery::Ping {}).unwrap();
+        let response: SpecialResponse = from_binary(&res).unwrap();
         assert_eq!(response.msg, "pong");
     }
 
     #[test]
     fn custom_query_execute_capitalize() {
-        let res = custom_query_execute(&CustomQuery::Capital {
+        let res = custom_query_execute(&SpecialQuery::Capital {
             text: "fOObaR".to_string(),
         })
         .unwrap();
-        let response: CustomResponse = from_binary(&res).unwrap();
+        let response: SpecialResponse = from_binary(&res).unwrap();
         assert_eq!(response.msg, "FOOBAR");
     }
 
     #[test]
     fn custom_querier() {
         let deps = mock_dependencies_with_custom_querier(20, &[]);
-        let req: QueryRequest<_> = CustomQuery::Capital {
+        let req: QueryRequest<_> = SpecialQuery::Capital {
             text: "food".to_string(),
         }
         .into();
-        let response: CustomResponse = deps.querier.custom_query(&req).unwrap();
+        let response: SpecialResponse = deps.querier.custom_query(&req).unwrap();
         assert_eq!(response.msg, "FOOD");
     }
 }